fix warnings & errors exposed by use of nvcc -x c++ compilation mode on gcc

+++ examples/monte_carlo_disjoint_sequences.cu
@@ -4,6 +4,7 @@
 #include <thrust/transform_reduce.h>
 
 #include <iostream>
+#include <cmath>
 
 // The technique demonstrated in the example monte_carlo.cu 
 // assigns an independently seeded random number generator to each
@@ -50,7 +51,7 @@ struct estimate_pi : public thrust::unary_function<unsigned int,float>
       float y = u01(rng);
 
       // measure distance from the origin
-      float dist = sqrtf(x*x + y*y);
+      float dist = std::sqrt(x*x + y*y);
 
       // add 1.0f if (u0,u1) is inside the quarter circle
       if(dist <= 1.0f)
